Updated booking test.
Moved booking list for user to customer class. updated bookingstest
