+++ src/abs/model/AbstractUser.java
@@ -14,8 +14,6 @@ public abstract class AbstractUser implements User {
 
 	private String password; // Hashed??
 
-	private List<Booking> bookings = new ArrayList<Booking>();
-
 	/**
 	 * @param name
 	 * @param email
@@ -45,27 +43,7 @@ public abstract class AbstractUser implements User {
 
 	@Override
 	public String toString() {
-		return "AbstractUser [name=" + name + ", email=" + email + ", password=" + password + ", bookings=" + bookings
-				+ "]";
-	}
-
-	public List<Booking> getBookings() {
-		return bookings;
-	}
-
-	public void setBookings(List<Booking> bookings) {
-		this.bookings = bookings;
-	}
-
-	// to add a booking the booking is added to the list in the user.
-	public boolean addBooking(Booking booking) {
-		// appends the booking time
-		this.bookings.add(booking);
-
-		// eventually we should sort the list after appending, or better yet
-		// insert into the list where it should actually go.
-
-		return true;
+		return "AbstractUser [name=" + name + ", email=" + email + ", password=" + password + "]";
 	}
 
 }
