+++ src/abs/model/Customer.java
@@ -1,5 +1,8 @@
 package abs.model;
 
+import java.util.ArrayList;
+import java.util.List;
+
 /**
  * 
  *
@@ -9,6 +12,7 @@ public class Customer extends AbstractUser {
 	private String address;
 
 	private String phone;
+	private List<Booking> bookings = new ArrayList<Booking>();
 
 	/**
 	 * @param name
@@ -35,6 +39,34 @@ public class Customer extends AbstractUser {
 	public String toString() {
 
 		return super.toString() + "Customer [address=" + address + ", phone=" + phone + "]";
+	}
+
+	public List<Booking> getBookings() {
+		return bookings;
+	}
+
+	public void setBookings(List<Booking> bookings) {
+		this.bookings = bookings;
+	}
+
+	// to add a booking the booking is added to the list in the user.
+	public boolean addBooking(Booking booking) {
+		// appends the booking time
+		// this.bookings.add(booking);
+
+		// eventually we should sort the list after appending, or better yet
+		// insert into the list where it should actually go.
+
+		return this.bookings.add(booking);
+	}
+
+	/**
+	 * @param booking
+	 * @return
+	 */
+	public boolean cancelBooking(Booking booking) {
+
+		return this.bookings.remove(booking);
 	}
 
 }
